refactor in order to move some stuff from readCluster to ZclCluster and create a zclClsuter helper
